fix: drop redundant i32 casts flagged by clippy in job.rs
exit_code and the 3621_i32 literal are already i32, so the intermediate
`as i32` before `as i64` was a no-op cast that clippy's
unnecessary_cast lint flagged on this PR (two unresolved review
threads from github-advanced-security).

+++ src/job.rs
@@ -665,7 +665,7 @@ unsafe fn child_error(
                 FmtArg::Str((pre) as *const ::core::ffi::c_char),
                 FmtArg::Str((nm) as *const ::core::ffi::c_char),
                 FmtArg::Str((f_name) as *const ::core::ffi::c_char),
-                FmtArg::Int((exit_code) as i32 as i64),
+                FmtArg::Int((exit_code) as i64),
                 FmtArg::Str((post) as *const ::core::ffi::c_char),
                 FmtArg::Str(smode_or_empty(smode).as_ptr()),
             ],
@@ -3095,7 +3095,7 @@ unsafe fn construct_command_argv_internal(
                     (b"src/job.c\0" as *const u8 as *const ::core::ffi::c_char)
                         as *const ::core::ffi::c_char,
                 ),
-                FmtArg::Int((3621_i32) as i32 as i64),
+                FmtArg::Int((3621_i32) as i64),
             ],
         );
     }
